requirements.txt
fastapi>=0.115,<1
uvicorn>=0.32,<1
sqlalchemy>=2.0,<3
alembic>=1.14,<2
pydantic>=2.10,<3
pytest>=8.3,<9
httpx>=0.27,<1
python-multipart>=0.0.20,<1
Pillow>=11.0,<12
